requirements.txt
langgraph==0.0.18
pandas>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pytest>=7.0.0